siteblog/requirements.txt
Django==3.2.6
django-allauth==0.53.1
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-debug-toolbar==3.2.2
django-environ==0.10.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
flake8==6.0.0
gunicorn==20.1.0
isort==5.12.0
Pillow==9.0.1
psycopg2==2.9.5
PyJWT==2.6.0
whitenoise==6.4.0
